Finished the core part of the game
I have finished the core part of the game. The game is playable now.
Now it's time to really make the UI better and cooler looking. I want
to add more picture and change the color scheme. I will work on the
layout and design but that will be a litle bit later. Now it is time to
rest.