Completed the rock paper scissors game
I went a bit all out on this one. I created three more pages. One for
the introduction to the game, one for when the player wins and one for
when the player loses. I also added a few images to bring the vision
alive. If I could change anything, that would be the size of the images
and making them flex to the screen more. Also, I would use different
fonts, sizes, and colors. I may revisit this project in the future once
I get a little bit more experience under my belt.