I couldn't help myself
I changed the way the UI to fit the photos and game output to fit the
screen better. This is so the player doesn't have to needlessly scroll.
I like how it came out. I may change the font of the game so it pops a
little more. Also, I changed the computer page so the title of the page
reflects that the computer won instead of saying the player won even
though they didn't.