Minor edits for now
I've changed the name of the orginial index.html file to match.html
along with the matchStyle.css file because in Vercel, the page was
opening to that instead, instead of the intro. I moved the match files
into a new folder called match. I also added the javas.js. I also
moved took the files from the intro folder, renamed them to index.html
style.css and moved them to the root folder. Later on I'll mess with
the css to make it look better, but for now just wanted to get something
up and running.