Updated the look of the overall game
Revisiting this project from a while ago, I've added a new font and
changed the look of the game. I also added a newimage on to the match
page. I've increased the font size for readability and also updated the
look of the buttons so they don't look flat. So far I like how this
turned out and my revisit this in the future and add more features.